[update] validate that cannot delete if category id is equals to 3

+++ app/Http/Controllers/Api/NotesController.php
@@ -85,6 +85,8 @@ class NotesController extends Controller
      */
     public function destroy(Note $note)
     {
+        abort_if($note->category_id == 3, 403, 'No tienes permiso para eliminar esta nota');
+
         $note->delete();
 
         return [
